requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
sqlalchemy[asyncio]==2.0.44
psycopg[binary]==3.2.12
alembic==1.18.4
pydantic==2.13.0
pydantic-settings==2.13.1
httpx==0.28.1
PyJWT==2.12.1
pwdlib[argon2]==0.3.0
argon2-cffi==25.1.0
python-multipart==0.0.26
email-validator==2.3.0
structlog==25.5.0
tenacity==9.1.2
cryptography==45.0.7
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
ruff==0.15.10